reorganize Manager/Runner methods
